+++ R/utilities.R
@@ -3,30 +3,93 @@ base_url <- function() {
   return("https://api.biorxiv.org")
 }
 
+# Build the query url for the content endpoint with a DOI
+build_content_doi_url <- function(server, doi) {
+  url <- paste0(base_url(), "/details/", server, "/", doi)
+  return(url)
+}
+
+# Build the query url for the content endpoint with 'from' and 'to' dates
+build_content_interval_url <- function(server, from, to, skip) {
+  # make sure 'skip' parameter is not given in scientific notation, ie. 10e5
+  skip <- format(skip, scientific = FALSE)
+  url <- paste0(base_url(), "/details/", server, "/", from, "/", to, "/", skip)
+  return(url)
+}
+
+# Build the query url for the publisher endpoint
+build_publisher_url <- function(prefix, from, to, skip) {
+  # make sure 'skip' parameter is not given in scientific notation, ie. 10e5
+  skip <- format(skip, scientific = FALSE)
+  url <- paste0(base_url(), "/publisher/", prefix, "/", from, "/", to, "/", skip)
+  return(url)
+}
+
+# Build the query url for the published endpoint
+build_published_url <- function(from, to, skip) {
+  # make sure 'skip' parameter is not given in scientific notation, ie. 10e5
+  skip <- format(skip, scientific = FALSE)
+  url <- paste0(base_url(), "/pub/", from, "/", to, "/", skip)
+  return(url)
+}
+
+# Check internet connection
+check_internet_connection <- function() {
+  if (curl::has_internet() == FALSE) {
+    stop("No internet connection detected. ",
+         "Please connect to the internet and try again.",
+         call. = F)
+  }
+}
+
 # Make query to API and return full response content
-fetch_content <- function(url) {
-  request <- httr::GET(url = url)
-  handle_response(request = request)
-  content <- httr::content(request, as = "parsed")
+do_query <- function(url) {
+  request <- httr::RETRY(verb = "GET",
+                         url = url,
+                         pause_base = 1,
+                         pause_cap = 60,
+                         times = 5,
+                         quiet = TRUE,
+                         httr::timeout(60))
+  content <- handle_response(request = request)
   return(content)
 }
 
-# Handle HTTP responses
+# Handle http responses
+# In addition to typical http responses, the following two cases may occur:
+# 1. A 200 status code may be received, but the response body contains a message
+# that the connection is refused (probably due to high server loads). In this
+# case an error is thrown and the action is stopped.
+# 2. A 200 status code is received, bu the response body contains a status
+# stating that no results were returned, either because no results are expected
+# or due to discrepancies between the expected and actual number of results
+# returned by the API, which interferes with paging functions. In this case a
+# null response is returned to be handled by the individual API endpoint functions.
 handle_response <- function(request) {
   if (request$status_code == 200) {
-    status <- unlist(httr::content(request, as = "parsed")$messages)["status"]
-    if (status != "ok") {
-      stop(status, call. = F)
+    if (check_refused_connection(request)) {
+      stop("Connection refused. The server may be experiencing high loads. ",
+           "Please try again later.", call. = F)
+    } else if (check_empty_response(request)) {
+      content <- NULL
+    } else {
+      content <- httr::content(request, as = "parsed")
     }
-  } else if (request$status_code == 404) {
-    stop("404: Page not found", call. = F)
-  } else if (request$status_code == 500) {
-    stop("500: Internal server error", call. = F)
-  } else if (request$status_code == 400) {
-    stop("400: Bad request", call. = F)
-  } else {
-    stop("Something went wrong", call. = F)
+    return(content)
   }
+  httr::stop_for_status(request)
+}
+
+# Check if a "Connection refused" message is returned
+check_refused_connection <- function(request) {
+  message <- httr::content(request, as = "text", encoding = "UTF-8")
+  return(message == "Error : (2002) Connection refused")
+}
+
+# Check if a "no posts found" status is returned
+check_empty_response <- function(request) {
+  status <- unlist(httr::content(request, as = "parsed")$messages)["status"]
+  return(status == "no posts found")
 }
 
 # Return the data in the requested format
@@ -124,3 +187,5 @@ set_column_types <- function(df) {
                                           to_date)
   return(df)
 }
+
+